core/godebug: make editor self contained to be able to run godebug without external dependencies.
Running godebug annotates source files. These annotations depend
on the "godebug/debug" package to communicate the values back to
the editor. The editor did not have this package (it is just a binary),
and so it needed to fetch that needed package from the web to be able
to compile programs with godebug annotations.

This patch solves this problem by including the "godebug/debug" package
in the form of a string inside the editor binary itself (like
a resource, code packed at "godebug/debugpack").
So when a program gets annotated, the string gets dumped in the final
source tree of the annotated program (and therefore no need to fetch
from the web).

This new mechanism (of not fetching the pkg from the web) means the
"godebug/debug" is not known by goimports. And so when editing a
program and adding "debug.AnnotateFile" in a line would not make
goimports insert the line "import ...godebug/debug" (major annoyance).

So the usage of "debug.AnnotateFile" func calls was moved to simply
using a comment in the form of "//godebug:annotatefile". This solves
the issue at a cost of a few cycles. Please refer to the README for
the annotation comments format.